Fix network

+++ include/utils.hh
@@ -34,7 +34,7 @@ class Utils
     // Network protocoled functions
 
     static uint64_t recvBytecodeLen(int socket);
-    static uint64_t sendBytecode(int socket, std::string& buffer, unsigned int len);
+    static uint64_t sendBytecode(int socket, std::string& buffer, uint64_t len);
 };
 
 #endif   /* !UTILS_HH */
